+++ aoindustries/net/net_devices-create.sql
@@ -17,11 +17,11 @@ create table net_devices (
   network text,
   broadcast text,
   mac_address text,
-  max_bit_rate int8,
-  monitoring_bit_rate_low int8,
-  monitoring_bit_rate_medium int8 check (monitoring_bit_rate_medium is null or monitoring_bit_rate_low is not null and monitoring_bit_rate_medium>monitoring_bit_rate_low),
-  monitoring_bit_rate_high int8 check (monitoring_bit_rate_high is null or monitoring_bit_rate_medium is not null and monitoring_bit_rate_high>monitoring_bit_rate_medium),
-  monitoring_bit_rate_critical int8 check (monitoring_bit_rate_critical is null or monitoring_bit_rate_high is not null and monitoring_bit_rate_critical>monitoring_bit_rate_high),
+  max_bit_rate bigint,
+  monitoring_bit_rate_low bigint,
+  monitoring_bit_rate_medium bigint check (monitoring_bit_rate_medium is null or monitoring_bit_rate_low is not null and monitoring_bit_rate_medium>monitoring_bit_rate_low),
+  monitoring_bit_rate_high bigint check (monitoring_bit_rate_high is null or monitoring_bit_rate_medium is not null and monitoring_bit_rate_high>monitoring_bit_rate_medium),
+  monitoring_bit_rate_critical bigint check (monitoring_bit_rate_critical is null or monitoring_bit_rate_high is not null and monitoring_bit_rate_critical>monitoring_bit_rate_high),
   monitoring_enabled boolean
     not null
     default true,
